ParabolicTrigonometry.jl

ParabolicTrigonometry.cosp — Method

cosp(ϕ)

The parabolic cosine function cosp, related to cos and cosh. See G. Dattoli, M. Migliorati, M. Quattromini, and P. E. Ricci, arXiv:1102.1563 [math-ph] for a definition.

The parabolic cosine function satisfies

cosp(0) = 1
sinp(0) = 0
cosp(ϕ)^2 + sin(ϕ) = 1
cosp(ϕ)' = - 1 / (sinp(ϕ) + 2 cosp(ϕ)^2)
sinp(ϕ)' = 2 cosp(ϕ) / (sinp(ϕ) + 2 cosp(ϕ)^2)

ParabolicTrigonometry.gudemann — Method

gudemann(ϕ)

The Gudemann function gd, defined by tan(gd(ϕ)) = tanh(ϕ).

ParabolicTrigonometry.sinp — Method

sinp(ϕ)

The parabolic sine function sinp, related to sin and sinh. See G. Dattoli, M. Migliorati, M. Quattromini, and P. E. Ricci, arXiv:1102.1563 [math-ph] for a definition.

The parabolic sine function satisfies

cosp(0) = 1
sinp(0) = 0
cosp(ϕ)^2 + sin(ϕ) = 1
cosp(ϕ)' = - 1 / (sinp(ϕ) + 2 cosp(ϕ)^2)
sinp(ϕ)' = 2 cosp(ϕ) / (sinp(ϕ) + 2 cosp(ϕ)^2)